# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server AI-HACKATHON.",
        "title": "AI HACKATHON 2025",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "ai-hackathon-2025-api.onrender.com",
    "basePath": "/",
    "paths": {
...
        "/v1/transactions": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get User profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/views.Transaction"
                            }
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't Get user",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get Result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Transaction data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/views.TransactionData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/views.ClassificationResult"
                        }
                    },
                    "401": {
                        "description": "No token Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found User Middleware",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Couldn't Get user",
                        "schema": {
                            "$ref": "#/definitions/views.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "views.ClassificationResult": {
            "type": "object",
            "properties": {
                "probability_of_fraud": {
                    "type": "number"
                }
            }
        },
...
        "views.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation_error"
                },
                "message": {
                    "type": "string",
                    "example": "Phone number already registered"
                }
            }
        },
...
        "views.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "destination": {
                    "type": "string"
                },
                "full_name_of_sender": {
                    "type": "string"
                },
                "target": {
                    "type": "integer"
                },
                "transdatetime": {
                    "type": "string"
                }
            }
        },
        "views.TransactionData": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "avg_login_interval_30d": {
                    "type": "number"
                },
                "burstiness_login_interval": {
                    "type": "number"
                },
                "ewm_login_interval_7d": {
                    "type": "number"
                },
                "fano_factor_login_interval": {
                    "type": "number"
                },
                "flogins_7d_over_30d_ratio": {
                    "type": "number"
                },
                "freq_change_7d_vs_mean": {
                    "type": "number"
                },
                "last_os_categorical": {
                    "type": "string"
                },
                "last_phone_model_categorical": {
                    "type": "string"
                },
                "login_frequency_7d": {
                    "type": "number"
                },
                "logins_last_30_days": {
                    "type": "number"
                },
                "logins_last_7_days": {
                    "type": "number"
                },
                "monthly_os_changes": {
                    "type": "integer"
                },
                "monthly_phone_model_changes": {
                    "type": "integer"
                },
                "std_login_interval_30d": {
                    "type": "number"
                },
                "transdate": {
                    "type": "string"
                },
                "var_login_interval_30d": {
                    "type": "number"
                },
                "zscore_avg_login_interval_7d": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}